requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==2.8
curl==0.0.1
daphne==2.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.41.0
django-braces==1.14.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-extensions==2.2.9
django-filter==2.2.0
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-registration==3.0.1
django-registration-redux==2.7
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-multiple-models==2.1.3
django-taggit==1.2.0
django-taggit-serializer==0.1.7
django-webpack-loader==0.6.0
djangorestframework==3.10.3
djangorestframework-api-key==2.0.0
djangorestframework-jwt==1.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.1.1
msgpack==0.6.2
oauthlib==3.1.0
Pillow==7.1.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyotp==2.3.0
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
stripe==2.48.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
whitenoise==5.0.1
zope.interface==4.7.1
django-webpack-loader
djangorestframework-api-key
django-rest-auth
django-taggit-serializer
django-taggit
oauth2_provider
social-auth-app-django
stripe
